More fixes

+++ 1D/Ladder.cpp
@@ -39,13 +39,13 @@ int main(int argc, char **argv) {
         if (X == 0)
             p = 1;
         else
-            p = (env[X - 1] == 1 ? 2.0 / 3.0 : 1.0 / 2.0);
+            p = (env[size_t(X - 1)] == 1 ? 2.0 / 3.0 : 1.0 / 2.0);
 
         int dX = vb::prng.bernoulli(p) ? 1 : -1;
 
         if ((X == int(env.size())) && (dX == 1)) env.push_back(2);
-        if ((dX == 1) && (env[X] == 1) && vb::prng.bernoulli(.5)) env[X] = 0;
-        if ((dX == 1) && (env[X] == 2)) env[X] = 1;
+        if ((dX == 1) && (env[size_t(X)] == 1) && vb::prng.bernoulli(.5)) env[size_t(X)] = 0;
+        if ((dX == 1) && (env[size_t(X)] == 2)) env[size_t(X)] = 1;
         X += dX;
     }
 }
